Fix: - Bug when created Entity from config can't created again

+++ Graph/Assets/Ruinum.ECS/Core/Configurations/Game/Strategies/Targets/ContextTargetStrategy.cs
@@ -17,7 +17,7 @@ namespace Ruinum.ECS.Configurations.Game.Strategies.Targets
                 if(Logging) LogErrorNotFound(nameof(config), (nameof(entity), entity));
                 return false;
             }
-            var entities = Contexts.game.GetEntitiesWithConfigIndex((uint)config.ConfigIndex);
+            var entities = Contexts.game.GetEntitiesWithConfigIndex(config.ConfigIndex);
             if (entities.Count == 0)
             {
                 targetEntity = default;
